+++ src/main/java/services/CommentsService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.math.BigInteger;
 
 @Service
 public class CommentsService {
@@ -15,10 +16,10 @@ public class CommentsService {
     PhotosService photosService;
 
     @Transactional
-    public void post(String content, int photo_id) {
+    public void post(String content, BigInteger photo_id) {
         commentDAO.saveComment(content, photo_id);
     }
-    public void deleteCommentByContent(String content,int id){
+    public void deleteCommentByContent(String content,BigInteger id){
         commentDAO.deleteCommentByPhotoIdandContent(content,id);
     }
 }
